+++ build-module/index.js
@@ -18,7 +18,7 @@ import { ShortcutProvider } from '@wordpress/keyboard-shortcuts';
 
 import BlockEditorContainer from './components/block-editor-container';
 import withRegistryProvider from './components/with-registry-provider';
-import EditorSetup from './components/editor-setup';
+import useEditorSetup from './components/editor-setup';
 import PatternMonitor from './components/pattern-monitor';
 import ContentSaver from './components/content-saver';
 import registerApiHandlers from './components/api-fetch';
@@ -124,6 +124,7 @@ import './style.scss';
  * @property {object[]} reusableBlocks
  * @property {object[]} styles
  * @property {object[]} defaultEditorStyles
+ * @property {string} bodyPlaceholder
  */
 
 /**
@@ -238,17 +239,18 @@ function IsolatedBlockEditor(props) {
     children,
     onSaveContent,
     onSaveBlocks,
-    settings,
     __experimentalUndoManager,
     __experimentalOnInput,
     __experimentalOnChange,
     __experimentalValue,
     __experimentalOnSelection,
     ...params
-  } = props;
+  } = props; // This needs to happen first to setup Gutenbergy things
+
   useInitializeIsoEditor({
     undoManager: __experimentalUndoManager
   });
+  const settings = useEditorSetup(props.settings);
   const editorSelection = useSelect(select => ({
     start: select('core/block-editor').getSelectionStart(),
     end: select('core/block-editor').getSelectionEnd()
@@ -259,8 +261,6 @@ function IsolatedBlockEditor(props) {
   return createElement(StrictMode, null, createElement(ShortcutProvider, null, createElement(ContentSaver, {
     onSaveBlocks: onSaveBlocks,
     onSaveContent: onSaveContent
-  }), createElement(EditorSetup, {
-    settings: settings
   }), createElement(PatternMonitor, null), createElement(SlotFillProvider, null, createElement(BlockEditorContainer, _extends({}, params, {
     onInput: __experimentalOnInput,
     onChange: __experimentalOnChange,