+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.22.0] - 2023-01-03
+
+### Added
+- Add __experimentalUndo for editing that better matches core Gutenberg
+- Replace refx with equivalent core version
+
 ## [2.21.0] - 2022-12-06
 
 ### Breaking changes